Reorganized sub-namespaces inside Traction.Contracts
