+++ Traction.CompileTimeDiagnosticsTests/InheritenceTests.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Microsoft.CodeAnalysis.CSharp;
 using NUnit.Framework;
-using static Traction.Contracts.Rewriting.DiagnosticCodes;
+using static Traction.Contracts.Analysis.DiagnosticCodes;
 using static Traction.Roslyn.Rewriting.DiagnosticCodes;
 
 namespace Traction.DiagnosticsTests {
